+++ TLEngine/renderMethods/TLRenderTemplate.mjs
@@ -1,0 +1,9 @@
+import fs from "node:fs";
+
+const TLRenderTemplate = (instructions) => {
+    const file = fs.readFileSync(instructions[0]);
+
+    return file.toString();
+}
+
+export { TLRenderTemplate }
